Implement vmm api adapters
The newly added module exposes vmm user-facing API and acts as an
adapter from the user-facing API to the vmm-as-a-library API.

This allows greater flexibility in changing the Vmm implementation
and code-level interface without breaking user-facing API.

Instead of trying to catch all the possible Vmm configuration
combinations and which APIs are allowed when, API Adapter flavors
also act as recipes for working with allowed subsets of the
user-facing API commands which are only relevant in certain usecases
(like preboot and postboot, snapshots and possibly live migration or
live update).
