+++ src/api_server/src/parsed_request.rs
@@ -18,7 +18,7 @@ use request::net::{parse_patch_net, parse_put_net};
 use request::vsock::parse_put_vsock;
 use ApiServer;
 
-use vmm::controller::{VmmAction, VmmActionError};
+use vmm::rpc_interface::{VmmAction, VmmActionError};
 
 #[allow(clippy::large_enum_variant)]
 pub enum ParsedRequest {
@@ -228,7 +228,7 @@ mod tests {
 
     use micro_http::HttpConnection;
     use vmm::builder::StartMicrovmError;
-    use vmm::controller::VmmActionError;
+    use vmm::rpc_interface::VmmActionError;
     use vmm::vmm_config::machine_config::VmConfig;
 
     impl PartialEq for ParsedRequest {
