+++ src/vmm/src/resources.rs
@@ -71,7 +71,7 @@ impl VmResources {
     /// Configures Vmm resources as described by the `config_json` param.
     pub fn from_json(
         config_json: &str,
-        firecracker_version: String,
+        firecracker_version: &str,
     ) -> std::result::Result<Self, Error> {
         let vmm_config: VmmConfig = serde_json::from_slice::<VmmConfig>(config_json.as_bytes())
             .map_err(|_| Error::InvalidJson)?;
